requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.37.0
Django==4.0.10
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pillow==10.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.42.0
